% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add Movebank data to a Frictionless Data Package}
\usage{
add_resource(package, resource_name, files, keys = TRUE)
}
\arguments{
\item{package}{List describing a Data Package, created with \code{\link[frictionless:read_package]{read_package()}}
or \code{\link[frictionless:create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{files}{One or more paths to CSV file(s) that contain the data for
this resource, as a character (vector).}

\item{keys}{Should \code{primaryKey} and \code{foreignKey} properties be added to the
Table Schema?}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
Adds Movebank data (\code{reference-data}, \code{gps}, \code{acceleration},
\code{accessory-measurements}) as a Data Resource to a Frictionless Data Package.
This function extends and masks \code{frictionless::add_resource()}.
}
\examples{
# See https://inbo.github.io/movepub/articles/movepub.html#frictionless
}
