% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite_to_eml.R
\name{datacite_to_eml}
\alias{datacite_to_eml}
\title{Get DataCite metadata as EML}
\usage{
datacite_to_eml(doi, contact = NULL, metadata_provider = contact)
}
\arguments{
\item{doi}{DOI of a dataset.}

\item{contact}{One or more persons to be set as resource contact, e.g.
\code{person("Peter", "Desmet", , "fakeaddress@email.com", "mdc", c(ORCID = "0000-0002-8442-8025"))}.}

\item{metadata_provider}{One or more persons to be set as metadata provider.
Same format as \code{contact}.}
}
\value{
EML list that can be extended and/or written to file with
\code{\link[EML:write_eml]{EML::write_eml()}}.
}
\description{
Get metadata from \href{https://datacite.org/}{DataCite} and transform to EML.
}
