% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aphia_id.R
\name{get_aphia_id}
\alias{get_aphia_id}
\title{Get WoRMS AphiaID from a taxonomic name}
\usage{
get_aphia_id(x)
}
\arguments{
\item{x}{A (vector with) taxonomic name(s).}
}
\value{
Data frame with \code{name}, \code{aphia_id}, \code{aphia_lsid} and \code{aphia_url}.
}
\description{
This function wraps \code{\link[worrms:wm_name2id]{worrms::wm_name2id_()}} so that it returns a data frame
rather than a list.
It also silences "not found" warnings, returning \code{NA} instead.
}
\examples{
get_aphia_id("Mola mola")
get_aphia_id(c("Mola mola", "not_a_name"))
}
\seealso{
Other support functions: 
\code{\link{datacite_to_eml}()}
}
\concept{support functions}
