% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvb_term.R
\name{get_mvb_term}
\alias{get_mvb_term}
\title{Get term from the Movebank Attribute Dictionary}
\usage{
get_mvb_term(label)
}
\arguments{
\item{label}{Label of the term to look for. Case will be ignored and \code{-},
\verb{_}, \code{.} and \code{:} interpreted as space.}
}
\value{
List with term information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated in favour of \code{\link[move2:movebank_get_vocabulary]{move2::movebank_get_vocabulary()}},
which offers the same functionality and more.

Search a term by its label in the \href{http://vocab.nerc.ac.uk/collection/MVB/current/}{Movebank Attribute Dictionary (MVB)}.
Returns in order: term with matching \code{prefLabel}, matching \code{altLabel} or
error when no matching term is found.
}
\examples{
get_mvb_term("animal_id")
get_mvb_term("individual-local-identifier") # A deprecated term
get_mvb_term("Deploy.On.Date")

# With move2
library(move2)
movebank_get_vocabulary("animal_id", return_type = "list")
movebank_get_vocabulary(
 "individual-local-identifier",
 omit_deprecated = TRUE,
 return_type = "list"
)
movebank_get_vocabulary("Deploy.On.Date", return_type = "list")
}
\keyword{internal}
