% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmtl_to_docbook.R
\name{html_to_docbook}
\alias{html_to_docbook}
\title{Convert HTML to DocBook XML}
\usage{
html_to_docbook(text)
}
\arguments{
\item{text}{String, may contain HTML.}
}
\value{
String, with HTML converted to DocBook XML.
}
\description{
Converts HTML strings to DocBook XML strings.
Handles the following tags: div, p, b, strong, i, em, a, ul, ol, li, sup,
sub, pre and h1:h6.
}
\section{Transformation details}{

\verb{<title>}
\itemize{
\item \verb{<h1>}
\verb{<para>}
\item \verb{<div>}
\item \verb{<p>}
\item \verb{<h2>}to \verb{<h6>}
}

\verb{<emphasis>}
\itemize{
\item \verb{<b>}
\item \verb{<strong>}
\item \verb{<i>}
\item \verb{<em>}
}

\verb{<ulink>}
\itemize{
\item \verb{<a href="...">}
}

\verb{<itemizedlist>}
\itemize{
\item \verb{<ul>}
}

\verb{<orderedlist>}
\itemize{
\item \verb{<ol>}
}

\verb{<listitem>}
\itemize{
\item \verb{<li>}
}

\verb{<superscript>}
\itemize{
\item \verb{<sup>}
}

\verb{<subscript>}
\itemize{
\item \verb{<sub>}
}

\verb{<literalLayout>}
\itemize{
\item \verb{<pre>}
}
}

\examples{
html_to_docbook("<div>text - <b>bold</b></div>")
}
\seealso{
Other support functions: 
\code{\link{datacite_to_eml}()},
\code{\link{get_aphia_id}()}
}
\concept{support functions}
