% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-frictionless.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{create_package}
\alias{read_package}
\alias{write_package}
\alias{resource_names}
\alias{read_resource}
\alias{remove_resource}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{frictionless}{\code{\link[frictionless:create_package]{create_package()}}, \code{\link[frictionless:read_package]{read_package()}}, \code{\link[frictionless:read_resource]{read_resource()}}, \code{\link[frictionless:remove_resource]{remove_resource()}}, \code{\link[frictionless:resource_names]{resource_names()}}, \code{\link[frictionless:write_package]{write_package()}}}
}}

