% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movebank_resource.R
\name{add_movebank_resource}
\alias{add_movebank_resource}
\title{Add Movebank Data Resource}
\usage{
add_movebank_resource(package, resource_name, files, keys = TRUE)
}
\arguments{
\item{package}{List describing a Data Package, created with \code{\link[frictionless:read_package]{read_package()}}
or \code{\link[frictionless:create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{files}{One or more paths to CSV file(s) that contain the data for
this resource, as a character (vector).}

\item{keys}{Should \code{primaryKey} and \code{foreignKey} properties be added to the
Table Schema?}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
\code{\link[frictionless:add_resource]{frictionless::add_resource()}} for \href{https://www.movebank.org/}{Movebank}
data.
}
