% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add Movebank data to a Frictionless Data Package}
\usage{
add_resource(package, resource_name, files, keys = TRUE)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[frictionless:read_package]{read_package()}} or
\code{\link[frictionless:create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{files}{One or more paths to CSV file(s) that contain the data for
this resource, as a character (vector).}

\item{keys}{If \code{TRUE}, \code{primaryKey} and \code{foreignKey} properties are added to
the Table Schema.}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
Adds Movebank data (\code{reference-data}, \code{gps}, \code{acceleration},
\code{accessory-measurements}) as a  Data Resource to a Frictionless Data Package.
The function extends \code{\link[frictionless:add_resource]{frictionless::add_resource()}}.
The title, definition, format and URI of each field are looked up in the
latest version of the \href{http://vocab.nerc.ac.uk/collection/MVB/current}{Movebank Attribute Dictionary} and included in the Table
Schema of the resource.
}
\details{
See \code{vignette("frictionless")} for examples.
}
\concept{frictionless functions}
