% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite_to_eml.R
\name{datacite_to_eml}
\alias{datacite_to_eml}
\title{Convert DataCite metadata to EML}
\usage{
datacite_to_eml(doi, contact = NULL, metadata_provider = contact)
}
\arguments{
\item{doi}{DOI of a dataset.}

\item{contact}{One or more persons to be set as resource contact. Provide as
person objects, e.g. \code{person("Peter", "Desmet", , "fakeaddress@email.com", "mdc", c(ORCID = "0000-0002-8442-8025"))}.}

\item{metadata_provider}{One or more person to be set as metadata provider.
Same format as \code{contact}.}
}
\value{
EML list that can be extended and/or written to file with
\code{\link[EML:write_eml]{EML::write_eml()}}.
}
\description{
Retrieves metadata from DataCite and converts to EML.
}
