% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite_to_eml.R
\name{datacite_to_eml}
\alias{datacite_to_eml}
\title{Get DataCite metadata as EML}
\usage{
datacite_to_eml(doi)
}
\arguments{
\item{doi}{DOI of a dataset.}
}
\value{
EML list that can be extended and/or written to file with
\code{\link[EML:write_eml]{EML::write_eml()}}.
}
\description{
Get metadata from \href{https://datacite.org/}{DataCite} and transform to EML.
}
\examples{
datacite_to_eml("10.5281/zenodo.10053903")

}
\seealso{
Other support functions: 
\code{\link{get_aphia_id}()},
\code{\link{html_to_docbook}()}
}
\concept{support functions}
