% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite_to_eml.R
\name{datacite_to_eml}
\alias{datacite_to_eml}
\title{Convert DataCite metadata to EML}
\usage{
datacite_to_eml(doi)
}
\arguments{
\item{doi}{DOI of a dataset.}
}
\value{
EML list that can be extended and/or written to file with
\code{\link[EML:write_eml]{EML::write_eml()}}.
}
\description{
Gets metadata for a DOI from DataCite and converts it to EML that can be
uploaded to a \href{https://www.gbif.org/ipt}{GBIF IPT}.
}
