% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_movebank_term.R
\name{get_movebank_term}
\alias{get_movebank_term}
\title{Get Movebank term}
\usage{
get_movebank_term(label)
}
\arguments{
\item{label}{Preferred label of the term.}
}
\value{
List object with term information.
}
\description{
Get information for a term in the \href{http://vocab.nerc.ac.uk/collection/MVB/current/}{Movebank Attribute Dictionary}.
}
\examples{
# Get information for term "animal ID"
get_movebank_term("animal ID")
}
