% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_movebank_term.R
\name{get_movebank_term}
\alias{get_movebank_term}
\title{Get Movebank term}
\usage{
get_movebank_term(label)
}
\arguments{
\item{label}{Label of the term to look for. Case will be ignored and \code{-},
\verb{_}, \code{.} and \code{:} interpreted as space.}
}
\value{
List object with term information.
}
\description{
Search a term by its label in the \href{http://vocab.nerc.ac.uk/collection/MVB/current/}{Movebank Attribute Dictionary}. Returns in
order: term with matching \code{prefLabel}, matching \code{altLabel} or error when no
matching term is found.
}
\examples{
get_movebank_term("animal_id")

get_movebank_term("Deploy.On.Date")
}
