% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvb_term.R
\name{get_mvb_term}
\alias{get_mvb_term}
\title{Get term from the Movebank Attribute Dictionary}
\usage{
get_mvb_term(label)
}
\arguments{
\item{label}{Label of the term to look for. Case will be ignored and \code{-},
\verb{_}, \code{.} and \code{:} interpreted as space.}
}
\value{
List with term information.
}
\description{
Search a term by its label in the \href{http://vocab.nerc.ac.uk/collection/MVB/current/}{Movebank Attribute Dictionary (MVB)}.
Returns in order: term with matching \code{prefLabel}, matching \code{altLabel} or
error when no matching term is found.
}
\examples{
get_mvb_term("animal_id")

get_mvb_term("Deploy.On.Date")
}
