% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_docbook.R
\name{html_to_docbook}
\alias{html_to_docbook}
\title{Convert text with HTML to DocBook}
\usage{
html_to_docbook(string)
}
\arguments{
\item{string}{Text that may contain HTML.}
}
\value{
Text with HTML converted to DocBook.
}
\description{
Converts HTML to \href{https://docbook.org/}{DocBook} for a given text.
Only a subset of HTML/DocBook tags are supported, see transformation details.
}
\section{Transformation details}{

The function only converts HTML tags that can be translated to DocBook tags
supported by EML for the \href{https://eml.ecoinformatics.org/schema/eml-text_xsd.html#TextType_para}{paragraph}
element.
The following replacements are made:\tabular{ll}{
   HTML tag \tab DocBook tag \cr
   \verb{<p>...</p>} \tab \verb{<para>...</para>} \cr
   \verb{<div>...</div>} \tab \verb{<para>...</para>} \cr
   \verb{<h2>...</h2>} \tab \verb{<para>...</para>} \cr
   \verb{<h3>...</h3>} \tab \verb{<para>...</para>} \cr
   \verb{<h4>...</h4>} \tab \verb{<para>...</para>} \cr
   \verb{<h5>...</h4>} \tab \verb{<para>...</para>} \cr
   \verb{<h6>...</h4>} \tab \verb{<para>...</para>} \cr
   \verb{<ul>...</ul>} \tab \verb{<itemizedlist>...</itemizedlist>} \cr
   \verb{<ol>...</ol>} \tab \verb{<orderedlist>...</orderedlist>} \cr
   \verb{<li>...</li>} \tab \verb{<listitem>...</listitem>} \cr
   \verb{<strong>...</strong>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<b>...</b>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<em>...</em>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<i>...</i>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<sub>...</sub>} \tab \verb{<subscript>...</subscript>} \cr
   \verb{<sup>...</sup>} \tab \verb{<superscript>...</superscript>} \cr
   \verb{<pre>...</pre>} \tab \verb{<literalLayout>...</literalLayout>} \cr
   \verb{<a href="http://example.com">...</a>} \tab \verb{<ulink url="https://example.com"><citetitle>...</citetitle></ulink>} \cr
   \verb{<h1>...</h1>} \tab \verb{<title>...</title>} \cr
}
}

\examples{
html_to_docbook("<div>My <b>bold</b> text.</div>")
}
\seealso{
Other support functions: 
\code{\link{datacite_to_eml}()},
\code{\link{get_aphia_id}()}
}
\concept{support functions}
