% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_docbook.R
\name{html_to_docbook}
\alias{html_to_docbook}
\title{Convert text with HTML to DocBook}
\usage{
html_to_docbook(string)
}
\arguments{
\item{string}{Character (vector) that may contain HTML syntax.}
}
\value{
A character vector with HTML converted to DocBook.
}
\description{
Converts text with HTML syntax to \href{https://docbook.org/}{DocBook}, splitting
paragraphs and headers into separate elements.
Only a subset of HTML tags are supported (see transformation details), all
other HTML syntax is removed.
}
\section{Transformation details}{

The function splits text into a character vector, with one element for each
paragraph, header or line break (\verb{\\n}).
The remaining HTML is converted to DocBook tags, but only those supported by
EML for \href{https://eml.ecoinformatics.org/schema/eml-text_xsd.html#TextType_para}{paragraphs}.
All other HTML/DocBook syntax is sanitized and empty elements are removed.\tabular{ll}{
   Input \tab Output \cr
   \verb{<h1>...</h1>} \tab \code{...} (separate element) \cr
   \verb{<p>...</p>} \tab \code{...} (separate element) \cr
   \verb{<div>...</div>} \tab \code{...} (separate element) \cr
   \verb{<h2>...</h2>} \tab \code{...} (separate element) \cr
   \verb{<h3>...</h3>} \tab \code{...} (separate element) \cr
   \verb{<h4>...</h4>} \tab \code{...} (separate element) \cr
   \verb{<h5>...</h4>} \tab \code{...} (separate element) \cr
   \verb{<h6>...</h4>} \tab \code{...} (separate element) \cr
   \verb{...\\n} \tab \code{...} (separate element) \cr
   \verb{<ul>...</ul>} \tab \verb{<itemizedlist>...</itemizedlist>} \cr
   \verb{<ol>...</ol>} \tab \verb{<orderedlist>...</orderedlist>} \cr
   \verb{<li>...</li>} \tab \verb{<listitem><para>...</para></listitem>} \cr
   \verb{<em>...</em>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<i>...</i>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<strong>...</strong>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<b>...</b>} \tab \verb{<emphasis>...</emphasis>} \cr
   \verb{<sub>...</sub>} \tab \verb{<subscript>...</subscript>} \cr
   \verb{<sup>...</sup>} \tab \verb{<superscript>...</superscript>} \cr
   \verb{<pre>...</pre>} \tab \verb{<literalLayout>...</literalLayout>} \cr
   \verb{<a href="http://example.com">...</a>} \tab \verb{<ulink url="https://example.com"><citetitle>...</citetitle></ulink>} \cr
   \verb{<code>...</code>} \tab \code{...} (HTML element sanitized) \cr
   \verb{<foo>...</foo>} \tab \code{...} (HTML element sanitized) \cr
   \verb{<span class="small">...</span>} \tab \code{...} (HTML property sanitized) \cr
   \verb{<p class="small">...</p>} \tab \code{...} (HTML property sanitized) \cr
   \verb{<img src="file.png">} \tab empty string HTML (element sanitized) \cr
   \verb{<emphasis>...</emphasis>} \tab \code{...} (DocBook element sanitized) \cr
}
}

\section{Use with EML}{

\enumerate{
\item Create or read EML with \code{EML::read_eml()}.
\item Assign output of \code{html_to_docbook()} to \code{eml$dataset$abstract$para}.
\item Write EML with \code{EML::write_eml()}.
}
}

\examples{
html_to_docbook(
  c(
    "This is <b>bold</b>.\nParagraph 1\n\nParagraph 2<p></p>",
    "What follows is a list: <ul><li>Item 1</li><li>Item 2</li></ul>"
  )
)
}
\seealso{
Other support functions: 
\code{\link{datacite_to_eml}()},
\code{\link{get_aphia_id}()}
}
\concept{support functions}
