% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmtl_to_docbook.R
\name{html_to_docbook}
\alias{html_to_docbook}
\title{Convert HTML to DocBook XML}
\usage{
html_to_docbook(text)
}
\arguments{
\item{text}{String, may contain HTML.}
}
\value{
String, with HTML converted to DocBook XML.
}
\description{
Converts HTML strings to DocBook XML strings.
Handles the following tags: div, p, b, strong, i, em, a, ul, ol, li, sup,
sub, h1, h2, h3, h4, h5, h6.
}
\examples{
html_to_docbook("<div>text - <b>bold</b></div>")
}
