% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepub-package.R
\docType{package}
\name{movepub-package}
\alias{movepub}
\alias{movepub-package}
\title{movepub: Prepare Movebank Data for Publication}
\description{
Prepare animal tracking data from 'Movebank' (\url{https://movebank.org}) for publication in a research repository. With 'movepub' you can document data with metadata following the Data Package standard and transform these to Darwin Core and Ecological Metadata Language ('EML') for publication to the Global Biodiversity Information Facility ('GBIF') and the Ocean Biodiversity Information System ('OBIS').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/movepub}
  \item \url{https://inbo.github.io/movepub}
  \item Report bugs at \url{https://github.com/inbo/movepub/issues}
}

}
\author{
\strong{Maintainer}: Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (Research Institute for Nature and Forest (INBO))

Other contributors:
\itemize{
  \item Sanne Govaert \email{sanne.govaert@inbo.be} (\href{https://orcid.org/0000-0002-8939-1305}{ORCID}) (Research Institute for Nature and Forest (INBO)) [contributor]
  \item Sarah Davidson \email{sdavidson@ab.mpg.de} (\href{https://orcid.org/0000-0002-2766-9201}{ORCID}) (Max Planck Institute of Animal Behavior) [contributor]
  \item Research Institute for Nature and Forest (INBO) (https://www.vlaanderen.be/inbo/en-gb/) [copyright holder]
  \item NLBIF (https://www.nlbif.nl/move2gbif-gps-zendergegevens-van-dieren-mobiliseren-naar-movebank-en-gbif/) [funder]
  \item European Union (https://dto-bioflow.eu/) [funder]
}

}
\keyword{internal}
