% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o_assen}
\alias{o_assen}
\title{Sample of a frictionless data package containing a bird tracking dataset}
\format{
An object of class \code{list} of length 4.
}
\source{
https://zenodo.org/record/5653311
}
\usage{
o_assen
}
\description{
A sample \href{https://specs.frictionlessdata.io/data-package/}{frictionless data package} of Movebank data, as read by \code{\link[frictionless:read_package]{frictionless::read_package()}}.
}
\details{
It is derived from a dataset on \href{https://zenodo.org/record/5653311}{Zenodo},
but excludes \verb{O_ASSEN-acceleration-2018.csv.gz} and \verb{O_ASSEN-acceleration-2019.csv.gz}.
}
\examples{
\dontrun{
# the data in o_assen was created with the code below
o_assen <- frictionless::remove_resource(
  frictionless::read_package("https://zenodo.org/record/5653311/files/datapackage.json"),
  "acceleration"
)
usethis::use_data(o_assen)
}
}
\concept{sample data}
\keyword{datasets}
