% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o_assen}
\alias{o_assen}
\title{Sample Movebank dataset with GPS tracking data}
\format{
An object of class \code{datapackage} (inherits from \code{list}) of length 6.
}
\source{
https://doi.org/10.5281/zenodo.10053903
}
\usage{
o_assen
}
\description{
A sample Movebank dataset with GPS tracking data, formatted as a
\href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package} and read by
\code{read_package()}.
}
\details{
This sample is derived from the Zenodo-deposited dataset
\href{https://doi.org/10.5281/zenodo.10053903}{Dijkstra et al. (2022)}, but
excludes the acceleration data.
}
\examples{
\dontrun{
# The data in o_assen was created with the code below
o_assen <-
  read_package("https://zenodo.org/records/10053903/files/datapackage.json") \%>\%
  remove_resource("acceleration")
o_assen$title <- "O_ASSEN - Eurasian oystercatchers (Haematopus ostralegus, Haematopodidae) breeding in Assen (the Netherlands)"
o_assen$licenses[[1]]$path <- "https://creativecommons.org/publicdomain/zero/1.0/"
usethis::use_data(o_assen)
}
}
\concept{sample data}
\keyword{datasets}
