% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-frictionless.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{create_package}
\alias{read_package}
\alias{write_package}
\alias{resources}
\alias{read_resource}
\alias{remove_resource}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{frictionless}{\code{\link[frictionless]{create_package}}, \code{\link[frictionless]{read_package}}, \code{\link[frictionless]{read_resource}}, \code{\link[frictionless]{remove_resource}}, \code{\link[frictionless]{resources}}, \code{\link[frictionless]{write_package}}}
}}

