% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_camtrap_dwca.R
\name{write_camtrap_dwca}
\alias{write_camtrap_dwca}
\title{Write camera trap Darwin Core Archive}
\usage{
write_camtrap_dwca(package)
}
\arguments{
\item{A}{Camera Trap Data Package.}
}
\value{
A Darwin Core Archive written to disk.
}
\description{
Write camera trap Darwin Core Archive
}
