% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Convert Movebank data to Darwin Core}
\usage{
write_dwc(
  package,
  directory = ".",
  doi = package$id,
  study_id,
  contact = NULL,
  metadata_provider = NULL,
  rights_holder = NULL
)
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.}

\item{directory}{Path to local directory to write files to.}

\item{doi}{DOI of the source dataset, used to populate record-level terms.}

\item{study_id}{Movebank study ID of the source dataset, used for reference.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{inheritParams}{datacite_to_eml}
}
\value{
CSV (data) and EML (metadata) files written to disk.
}
\description{
Converts a published Movebank dataset (formatted as a Frictionless Data
Package) to Darwin Core (CSV files) and EML (\code{eml.xml}).
The resulting files can be uploaded to a \href{https://www.gbif.org/ipt}{GBIF IPT}
for publication.
A \code{meta.xml} file is not created.
}
\details{
The conversion to Darwin Core is expressed in SQL:
\itemize{
\item \href{https://github.com/inbo/movepub/blob/main/inst/sql/movebank_dwc_occurrence.sql}{dwc_occurrence}
}
}
