% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Movebank data to Darwin Core}
\usage{
write_dwc(
  package,
  directory = ".",
  doi = package$id,
  contact = NULL,
  rights_holder = NULL,
  study_id = NULL
)
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.}

\item{directory}{Path to local directory to write files to.}

\item{doi}{DOI of the original dataset, used to get metadata.}

\item{contact}{Person to be set as resource contact and metadata provider.
To be provided as a \code{person()}.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{study_id}{Identifier of the Movebank study from which the dataset was
derived (e.g. \code{1605797471} for
\href{https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study160579747}{this study}).}
}
\value{
CSV (data) and EML (metadata) files written to disk.
}
\description{
Transforms a published Movebank dataset (formatted as a
\href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package})
to Darwin Core CSV and EML files that can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT} for publication.
A \code{meta.xml} file is not created.
}
\section{Metadata}{


Metadata are derived from the original dataset by looking up its \code{doi} in
DataCite (\href{https://api.datacite.org/dois/10.5281/zenodo.5879096}{example})
and transforming these to EML.
Uses \code{datacite_to_eml()} under the hood.
The following properties are set:
\itemize{
\item \strong{title}: Original title + \verb{[subsampled representation]}.
\item \strong{description}: Automatically created first paragraph describing this is
a derived dataset, followed by the original dataset description.
\item \strong{license}: License of the original dataset.
\item \strong{creators}: Creators of the original dataset.
\item \strong{contact}: \code{contact} or first creator of the original dataset.
\item \strong{metadata provider}: \code{contact} or first creator of the original dataset.
\item \strong{keywords}: Keywords of the original dataset.
\item \strong{alternative identifier}: DOI of the original dataset. This way, no new
DOI will be created when publishing to GBIF.
\item \strong{external link} and \strong{alternative identifier}: URL created from
\code{study_id} or the first "derived from" related identifier in the original
dataset.
}

To be set manually in the GBIF IPT: \strong{type}, \strong{subtype},
\strong{update frequency}, and \strong{publishing organization}.

Not set: geographic, taxonomic, temporal coverage, associated parties,
project data, sampling methods, and citations.
Not applicable: collection data.
}

\section{Data}{


\code{package} is expected to contain a \code{reference-data} and \code{gps} resource.
Their CSV data are loaded in to a SQLite database,
\href{https://github.com/inbo/movepub/blob/main/inst/sql/dwc_occurrence.sql}{transformed to Darwin Core using SQL}
and written to disk as CSV file(s).

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (animal+tag associations) are parent events, with tag
attachment (a human observation) and GPS positions (machine observations)
as child events.
No information about the parent event is provided other than its ID,
meaning that data can be expressed in an Occurrence Core with one row per
observation and \code{parentEventID} shared by all occurrences in a deployment.
\item The tag attachment event often contains metadata about the animal (sex,
lifestage, comments) and deployment as a whole.
\item No event/occurrence is created for the deployment end, since the end date
is often undefined, unreliable and/or does not represent an animal
occurrence.
\item Only \code{visible} (nonoutlier) GPS records that fall within a deployment are
included.
\item GPS positions are downsampled to the first GPS position per hour, to reduce
the size of high-frequency data.
It is possible for a deployment to contain no GPS positions, e.g. if the
tag malfunctioned right after deployment.
}
}

\examples{
# See https://inbo.github.io/movepub/articles/movepub.html#dwc
}
\concept{dwc functions}
