% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Movebank data to Darwin Core}
\usage{
write_dwc(package, directory = ".", doi = package$id, rights_holder = NULL)
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.}

\item{directory}{Path to local directory to write file to.}

\item{doi}{DOI of the original dataset, used to get dataset-level terms.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}
}
\value{
CSV (data) file written to disk.
}
\description{
Transforms data from a Movebank dataset (formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}) to \href{https://dwc.tdwg.org/}{Darwin Core}.
The resulting CSV file can be uploaded to an \href{https://www.gbif.org/ipt}{IPT}
for publication to GBIF and/or OBIS, together with an EML (metadata) file
created with \code{write_eml()}.
A \code{meta.xml} file is not created.
}
\details{
See \href{https://inbo.github.io/movepub/articles/movepub.html#dwc}{Get started}
for examples.
}
\section{Data}{

\code{package} is expected to contain a \code{reference-data} and \code{gps} resource.
Data are transformed into an \href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence Core}.
This \strong{follows recommendations} discussed and created by Peter Desmet,
Sarah Davidson, John Wieczorek and others.

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (animal+tag associations) are parent events, with tag
attachment (a human observation) and GPS positions (machine observations)
as child events.
No information about the parent event is provided other than its ID,
meaning that data can be expressed in an Occurrence Core with one row per
observation and \code{parentEventID} shared by all occurrences in a deployment.
\item The tag attachment event often contains metadata about the animal (sex,
lifestage, comments) and deployment as a whole.
\item No event/occurrence is created for the deployment end, since the end date
is often undefined, unreliable and/or does not represent an animal
occurrence.
\item Only \code{visible} (nonoutlier) GPS records that fall within a deployment are
included.
\item GPS positions are downsampled to the \strong{first GPS position per hour}, to
reduce the size of high-frequency data.
It is possible for a deployment to contain no GPS positions, e.g. if the
tag malfunctioned right after deployment.
}
}

\examples{
\dontrun{
write_dwc(o_assen)
}
}
\seealso{
Other dwc functions: 
\code{\link{write_eml}()}
}
\concept{dwc functions}
