% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Movebank data to a Darwin Core Archive}
\usage{
write_dwc(
  package,
  directory,
  dataset_id = package$id,
  dataset_name = package$title,
  license = NULL,
  rights_holder = NULL
)
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as returned by
\code{\link[=read_package]{read_package()}}.
It is expected to contain a \code{reference-data} and \code{gps} resource.}

\item{directory}{Path to local directory to write files to.}

\item{dataset_id}{Identifier for the dataset.}

\item{dataset_name}{Title of the dataset.}

\item{license}{License of the dataset.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}
}
\value{
CSV and \code{meta.xml} files written to disk.
And invisibly, a list of data frames with the transformed data.
}
\description{
Transforms a Movebank dataset (formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}) to a \href{https://dwc.tdwg.org/text/}{Darwin Core Archive}.
}
\details{
The resulting files can be uploaded to an \href{https://www.gbif.org/ipt}{IPT}
for publication to GBIF and/or OBIS.
A corresponding \code{eml.xml} metadata file can be created with \code{\link[=write_eml]{write_eml()}}.
See \code{vignette("movepub")} for an example.
}
\section{Transformation details}{

This function \strong{follows recommendations} suggested by Peter Desmet,
Sarah Davidson, John Wieczorek and others and transforms data to:
\itemize{
\item An \href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence core}.
\item An \href{https://rs.gbif.org/extension/obis/extended_measurement_or_fact_2023-08-28.xml}{Extended Measurements Or Facts extension}
\item A \code{meta.xml} file.
}

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (animal+tag associations) are parent events, with tag
attachment (a human observation) and GPS positions (machine observations)
as child events.
No information about the parent event is provided other than its ID,
meaning that data can be expressed in an Occurrence core with one row per
observation and \code{parentEventID} shared by all occurrences in a deployment.
\item The tag attachment event often contains metadata about the animal (sex,
life stage, comments) and deployment as a whole.
The sex and life stage are additionally provided in an Extended Measurement
Or Facts extension, where values are mapped to a controlled vocabulary
recommended by \href{https://obis.org/}{OBIS}.
\item No event/occurrence is created for the deployment end, since the end date
is often undefined, unreliable and/or does not represent an animal
occurrence.
\item Only \code{visible} (non-outlier) GPS records that fall within a deployment are
included.
\item GPS positions are downsampled to the \strong{first GPS position per hour}, to
reduce the size of high-frequency data.
It is possible for a deployment to contain no GPS positions, e.g. if the
tag malfunctioned right after deployment.
\item Parameters or metadata are used to set the following record-level terms:
\itemize{
\item \code{dwc:datasetID}: \code{dataset_id}, defaulting to \code{package$id}.
\item \code{dwc:datasetName}: \code{dataset_name}, defaulting to \code{package$title}.
\item \code{dcterms:license}: \code{license}, defaulting to the first license \code{name}
(e.g. \code{CC0-1.0}) in \code{package$licenses}.
\item \code{dcterms:rightsHolder}: \code{rights_holder}, defaulting to the first
contributor in \code{package$contributors} with role \code{rightsHolder}.
}
}
}

\examples{
write_dwc(o_assen, directory = "my_directory")

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
\seealso{
Other dwc functions: 
\code{\link{write_eml}()}
}
\concept{dwc functions}
