% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{Transform Movebank data to EML (metadata)}
\usage{
write_eml(
  package,
  directory = ".",
  doi = package$id,
  contact = NULL,
  rights_holder = NULL,
  study_id = NULL
)
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.}

\item{directory}{Path to local directory to write file(s) to.}

\item{doi}{DOI of the original dataset, used to get metadata.}

\item{contact}{Person to be set as resource contact and metadata provider.
To be provided as a \code{\link[=person]{person()}}.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{study_id}{Identifier of the Movebank study from which the dataset was
derived (e.g. \code{1605797471} for
\href{https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study160579747}{this study}).}
}
\value{
EML (metadata) file written to disk.
}
\description{
Transforms data from a Movebank dataset (formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}) to Ecological
Metadata Language (EML)
The resulting EML file can be uploaded to an \href{https://www.gbif.org/ipt}{IPT}
for publication to GBIF and/or OBIS, together with a CSV (data) file created
with \code{write_dwc()}
A \code{meta.xml} file is not created.
}
\details{
See \href{https://inbo.github.io/movepub/articles/movepub.html#dwc}{Get started}
for examples.
}
\section{Metadata}{

Metadata are derived from the original dataset by looking up its \code{doi} in
DataCite (\href{https://api.datacite.org/dois/10.5281/zenodo.5879096}{example})
and transforming these to EML.
Uses \code{datacite_to_eml()} under the hood.
The following properties are set:
\itemize{
\item \strong{title}: Original title + \verb{[subsampled representation]}.
\item \strong{description}: Automatically created first paragraph describing this is
a derived dataset, followed by the original dataset description.
\item \strong{license}: License of the original dataset.
\item \strong{creators}: Creators of the original dataset.
\item \strong{contact}: \code{contact} or first creator of the original dataset.
\item \strong{metadata provider}: \code{contact} or first creator of the original dataset.
\item \strong{keywords}: Keywords of the original dataset.
\item \strong{alternative identifier}: DOI of the original dataset. This way, no new
DOI will be created when publishing to GBIF.
\item \strong{external link} and \strong{alternative identifier}: URL created from
\code{study_id} or the first "derived from" related identifier in the original
dataset.
}

To be set manually in the GBIF IPT: \strong{type}, \strong{subtype},
\strong{update frequency}, and \strong{publishing organization}.

Not set: geographic, taxonomic, temporal coverage, associated parties,
project data, sampling methods, and citations.
Not applicable: collection data.
}

\examples{
\dontrun{
write_eml(o_assen)
}
}
\seealso{
Other dwc functions: 
\code{\link{write_dwc}()}
}
\concept{dwc functions}
