% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_movebank_dwca.R
\name{write_movebank_dwca}
\alias{write_movebank_dwca}
\title{Write Movebank Darwin Core Archive}
\usage{
write_movebank_dwca(package, directory = ".")
}
\arguments{
\item{package}{A Frictionless Data Package of Movebank data, as read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.}

\item{directory}{Path to local directory to write files to.}
}
\value{
Darwin Core Archive formatted CSV files written to disk.
}
\description{
Converts a Frictionless Data Package of Movebank data to Darwin Core Archive
formatted CSV files that can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT}.
The conversion is expressed in SQL:
\itemize{
\item \href{https://github.com/inbo/movepub/blob/main/inst/sql/movebank_dwc_occurrence_deployment.sql}{dwc_occurrence_deployment}
\item \href{https://github.com/inbo/movepub/blob/main/inst/sql/movebank_dwc_occurrence_positions.sql}{dwc_occurrence_positions}
}
}
